// src/store/chats/models/system-message-type.ts
export enum SystemMessageType {
  None = 'None',
  GroupChatMemberRemoved = 'GroupChatMemberRemoved',
  GroupChatAvatarChanged = 'GroupChatAvatarChanged',
  GroupChatCreated = 'GroupChatCreated',
  GroupChatMemberAdded = 'GroupChatMemberAdded',
  GroupChatNameChanged = 'GroupChatNameChanged',
  GroupChatAvatarRemoved = 'GroupChatAvatarRemoved',
  UserCreated = 'UserCreated',
  CallEnded = 'CallEnded',
}